Allow using a plain old validation function
