+++ src/index.ts
@@ -6,5 +6,6 @@ export * from "./useIdentifier";
 export * from "./useInsertItem";
 export * from "./usePushItem";
 export * from "./useRemoveItem";
+export * from "./useValidation";
 export * from "./useValidate";
 export * from "./useSubmit";
